Backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
elevenlabs
python-multipart
SpeechRecognition
pydub